make it work with pymc 5.x
